+++ liorm/__init__.py
@@ -2,8 +2,8 @@
 
 from liorm.capabilities import Profile, RawSqlCapability, assert_capability
 from liorm.catalog import CATALOG_ALLOWLIST
-from liorm.errors import CatalogError
-from liorm.execute import OrmError, execute, register_plan
+from liorm.errors import CatalogError, OrmError
+from liorm.execute import execute, register_plan
 
 __all__ = [
     "CATALOG_ALLOWLIST",